Remove a simple warning

+++ Source/DataStructureAndEncodingDefinition/gdcmVM.cxx
@@ -328,7 +328,7 @@ VM::VMType VM::GetVMTypeFromLength(size_t length, unsigned int size)
 {
   // Check first of length is a indeed a multiple of size and that length is != 0
   if ( !length || length % size ) return VM::VM0;
-  const unsigned int ratio = length / size;
+  const unsigned int ratio = (unsigned int)(length / size);
   //std::cerr << "RATIO=" << ratio << std::endl;
   switch( ratio )
     {
